Finished day 23. What a mission.
After a loooooong search, ended up using `dijkstra` from `Algorithm.Search` which happens to be exacted what I needed - it constructs the graph as it goes, unlike FGL which requires a fully constructed graph.

Had a long time latent bug in `pathsToHallway` that was enough to blow out run times but still return the right result, some times. I remember now it was a mental TODO that I never TODID. With that fixed, and a lot of repetitive code to handle Part 2 variations, it gets the job done in 19 seconds.